move k8s test to a //go:build k8s test
