+++ src/server/debug/server/debugstar/debugstar_test.go
@@ -74,23 +74,6 @@ func TestStarlark(t *testing.T) {
 	}
 }
 
-func TestBuiltinScripts(t *testing.T) {
-	var ran bool
-	for name, script := range BuiltinScripts {
-		t.Run(name, func(t *testing.T) {
-			ctx := pctx.TestContext(t)
-			env := &Env{FS: testDumpFS(fstest.MapFS{})}
-			if err := env.RunStarlark(ctx, script, string(script)); err != nil {
-				t.Errorf("run script: %v", err)
-			}
-			ran = true
-		})
-	}
-	if !ran {
-		t.Fatal("expected to run at least 1 built-in script")
-	}
-}
-
 func TestInteractiveDumpFS(t *testing.T) {
 	base := t.TempDir()
 	fs := &InteractiveDumpFS{Base: base}
